requirements.txt
requests>=2.25.1
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.4.1